feat: 그룹 생성 API연결 S12P31D107-144

See merge request s12-final/S12P31D107!63

+++ RunMate/data/src/main/java/com/D107/runmate/data/remote/response/kakaolocal/PlaceResponse.kt
@@ -30,8 +30,8 @@ data class PlaceResponse(
                 id = id,
                 name = placeName,
                 address = addressName,
-                x = x,
-                y = y
+                x = x.toDouble(),
+                y = y.toDouble()
             )
         }
